Trunk: HILSim compiler warning fixes, non-functional change. - Fixed compile warnings in VS2k8 where a float multiplication was elevated to double then reassigned to a float without the proper casting back to float. - Fixed compile warnings in VS2k8 where a bool was being assigned a large number hopefully treating it as true. - removed an unused variable which were causing a warning
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@2003 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ Tools/HILSIM_XPlane/utility.cpp
@@ -79,14 +79,14 @@ void BCBFtoOGL(float &x, float &y, float &z, float phi, float theta, float psi)
 
 		float tempx = ( x * Cp * Cy ) + ( y * (( Cy * Sr * Sp ) - ( Cr * Sy ))) + ( z * (( Cr * Cy * Sp ) + ( Sr * Sy )));
 		float tempy = ( x * Cp * Sy ) + ( y * (( Cr * Cy ) + ( Sr * Sp * Sy ))) + ( z * (( Cr * Sp * Sy ) - ( Cy * Sr )));
-		float tempz = ( x * Sp * -1.0) + ( y * Cp * Sr ) + ( z * Cr * Cp);
+		float tempz = (float)(( x * Sp * -1.0) + ( y * Cp * Sr ) + ( z * Cr * Cp));
 
 		// tempx, y & z should be in the NED frame, as that is where our roll, pitch, yaw angles are defined.
 		// need to convert them to East Up South, which is the OGL frame.
 
-		z = tempx * -1.0;	// tempx points north, z in OGL is +ve south
+		z = (float)(tempx * -1.0);	// tempx points north, z in OGL is +ve south
 		x = tempy;			// tempy points east, x in OGL is +ve east
-		y = tempz * -1.0;	// tempz points down, y in OGL is +ve up
+		y = (float)(tempz * -1.0);	// tempz points down, y in OGL is +ve up
 
 		return;
 }
@@ -106,9 +106,9 @@ void OGLtoBCBF(float &x, float &y, float &z, float phi, float theta, float psi)
 	//	The +Y axis points straight up away from the center of the earth at the reference point.
 	
 	// First we shall convert from this East Up South frame, to a more conventional NED (North East Down) frame.
-	x_NED = -1.0 * z;
+	x_NED = (float)(-1.0 * z);
 	y_NED = x;
-	z_NED = -1.0 * y; 
+	z_NED = (float)(-1.0 * y); 
 
 	// Next calculate cos & sin of angles for use in the transformation matrix.
 	// r, p & y subscripts stand for roll pitch and yaw.
